@@ -3,14 +3,9 @@ package com.pluralsight.blog.model;
 import org.hibernate.annotations.CreationTimestamp;
 
 import javax.persistence.*;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
-import java.lang.reflect.Method;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
 import java.util.Date;
-import java.util.List;
 
 @Entity
 public class Post {
@@ -74,9 +69,9 @@ public class Post {
 
     @Override
     public boolean equals(Object obj) {
-        if (!(obj instanceof Post))
+        if (!(obj instanceof Post otherPost))
             return false;
-        Post otherPost = (Post)obj;
+
         return this.title.equals(otherPost.getTitle()) &&
                this.body.equals(otherPost.getBody());
     }
